Move shallow_copy, deep_copy and add string utility functions.
