+++ scripts/utilities.ttslua
@@ -262,28 +262,6 @@ function corners_to_str(corner)
         'botleft='  .. vec_to_str(corner['botleft']) .. '}'
 end
 
-function shallow_copy(orig)
-  if not (type(orig) == 'table') then
-    return orig
-  end
-  local result={}
-  for k,v in pairs(orig) do
-    result[k] = v
-  end
-  return result
-end
-
-function deep_copy(orig)
-  if not (type(orig) == 'table') then
-    return orig
-  end
-  local result={}
-  for k,v in pairs(orig) do
-    result[k] = deep_copy(v)
-  end
-  return result
-end
-
 function get_any_key(a_table)
   for k,_ in pairs(a_table) do
     return k
